♻️  Refactoring(llm): move around class/defs/files
