+++ fai-rag-app/fai-backend/fai_backend/llm/models.py
@@ -1,0 +1,10 @@
+from datetime import datetime
+
+from pydantic import BaseModel
+
+
+class LLMMessage(BaseModel):
+    type: str
+    date: datetime
+    source: str | None = None
+    content: str | None = None
