+++ fai-rag-app/fai-backend/fai_backend/llm/serializer/impl/json.py
@@ -1,0 +1,7 @@
+from fai_backend.llm.models import LLMMessage
+from fai_backend.llm.serializer.protocol import ISerializer
+
+
+class JSONSerializer(ISerializer):
+    def serialize(self, input_data: LLMMessage) -> str:
+        return input_data.model_dump_json(exclude_none=True)
