+++ fai-rag-app/fai-backend/fai_backend/llm/serializer/protocol.py
@@ -1,0 +1,11 @@
+from typing import Protocol
+
+from fai_backend.llm.models import LLMMessage
+
+
+class ISerializer(Protocol):
+    def serialize(self, input_data: LLMMessage) -> str:
+        """
+
+        """
+        ...
